requirements.txt
anthropic>=0.46.0
python-dotenv>=1.0.0
